# Parallel calls for yt playlists