# 🚀 Pull Request

## 📋 Summary
# YouTube Playlist API Performance Optimization

## Problem
The YouTube playlist endpoint was making sequential API calls to fetch video details, causing slower response times especially for larger playlists. Each batch of 50 videos was being processed one after another, leading to unnecessary wait times.

## Changes Made
1. Added parallelization to video detail fetching
   - Gather all playlist items first
   - Create batched tasks for video details (50 videos per batch)
   - Use `Task.WhenAll()` to process batches concurrently
   - Combine results into a single dictionary

2. Added performance monitoring
   - Implemented Stopwatch to measure execution time
   - Added debug-level logging to track timing metrics
   - Ensures logging only occurs in development environment

## Performance Testing Results

### Test Case 1: Travis Scott Playlist
- Playlist ID: PLMtSW7uUwiQ33ky2J2880iXx6soVrqir7
- Video Count: 26 videos
- Before: 1005ms (≈1.0 seconds)
- After: 511ms (≈0.5 seconds)
- Improvement: ~49% faster

### Test Case 2: "Important Videos" Playlist
- Playlist ID: PLXKAG8g1Ls_Ax-SU7rCgyiGWjylB5NHL-
- Video Count: 103 videos
- Before: 2646ms (≈2.65 seconds)
- After: 1278ms (≈1.28 seconds)
- Improvement: ~52% faster

## 🔍 Related Issues
<!-- Link to related issues, e.g., "Fixes #123" or "Closes #456" -->
Fixes #9 

## 📷 Screenshots (if applicable)
<!-- Add screenshots or GIFs if visual changes were made. -->

## 💬 Additional Comments
- Maintained all existing functionality including shuffle feature
- Preserved YouTube API batch limit of 50 videos per request
- Added proper error handling for duration parsing
- Implemented debug-level logging to avoid production log clutter